requirements.txt
onnxruntime>=1.16.0
numpy>=1.24.0
soundfile>=0.12.1
pyyaml>=6.0.1
music21>=9.1.0
g2p_en>=2.1.0
RestrictedPython>=7.0
fastapi>=0.110.0
email-validator>=2.1.1
uvicorn>=0.27.0
python-multipart>=0.0.9
httpx>=0.27.0
ffmpeg-python>=0.2.0
firebase-admin>=6.5.0
google-cloud-storage>=2.14.0
google-cloud-secret-manager>=2.16.4
stripe>=12.0.0
